% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPGA_tablization.R
\name{BPGA_tablization}
\alias{BPGA_tablization}
\title{BPGA tablization}
\usage{
BPGA_tablization(supporting_files_dir = NULL)
}
\arguments{
\item{supporting_files_dir}{Path to the directory containing Supporting_files. Defaults to the current working directory.}
}
\value{
A reconstructed data frame.
}
\description{
BPGA tablization
}
