% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genbank_info_combiner.R
\name{Genbank_info_combiner}
\alias{Genbank_info_combiner}
\title{Combine GenBank Genome Information}
\usage{
Genbank_info_combiner(directory = NULL, save_output = FALSE)
}
\arguments{
\item{directory}{A character string specifying the path to the directory containing GenBank files.
If \code{NULL}, the current working directory (\code{getwd()}) or a subdirectory named
"genbank" will be used.}

\item{save_output}{Logical, whether to save outputs for each processed file. Default is \code{FALSE}.}
}
\value{
A data frame \code{Genome_summary_df} containing combined genome information from all processed GenBank files.
}
\description{
This function processes multiple GenBank files from a specified directory, extracts genome information
using the \code{gb_info} function, and compiles the results into a summary data frame. The processed data
are saved in the global environment as \code{Genome_summary_list} and \code{Genome_summary_df}.
}
\details{
The function:
\itemize{
\item Searches the specified directory for GenBank files with extensions \code{.gbk}, \code{.gb}, or \code{.gbff}.
\item Calls the \code{gb_info} function to process each GenBank file and extract genome information.
\item Compiles the results into a list (\code{Genome_summary_list}) and a flattened data frame (\code{Genome_summary_df}).
\item Saves both the list and data frame in the global environment.
}

If the specified directory does not contain GenBank files, or if the directory is invalid, the function will stop with an error.
}
\examples{
# Example usage
combine_gb_info(directory = "path/to/genbank/files")

# Automatically detect the 'genbank' folder or use the current working directory
combine_gb_info()

}
\seealso{
\code{\link{gb_info}}, \code{\link{Genbank_organizer_batch}}
}
