% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pan_core_plot.R
\name{pan_core_plot}
\alias{pan_core_plot}
\title{Generate Pan-Core Genome Plot}
\usage{
pan_core_plot(directory = NULL, save_plot = TRUE, plot_filename = NULL)
}
\arguments{
\item{directory}{The directory containing the required files: \code{curve.xls}, \code{list}, \code{pan_genome.txt},
and \code{core_genome.txt}. Defaults to the current working directory if \code{NULL}.}

\item{save_plot}{Logical value indicating whether to save the plot as a PDF. Defaults to \code{TRUE}.}

\item{plot_filename}{The name of the output PDF file. If \code{NULL}, defaults to \code{"pan_core_plot.pdf"}.}
}
\value{
A \code{ggplot} object representing the pan-core genome plot.
}
\description{
This function generates a pan-core genome plot based on provided genomic data files. The plot includes both
pan-genome and core-genome trends, along with associated boxplots and jitter plots for visualization.
}
\examples{
# Generate and save the plot in the current directory
pan_core_plot()

# Generate the plot without saving
pan_core_plot(save_plot = FALSE)

# Generate the plot and save to a specific directory
pan_core_plot(directory = "path/to/data", plot_filename = "custom_plot.pdf")
}
