% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanDB_generator.R
\name{PanDB_generator}
\alias{PanDB_generator}
\title{PanDB Generator}
\usage{
PanDB_generator(directory = NULL, save_output = FALSE)
}
\arguments{
\item{directory}{A character string specifying the path to the directory containing GenBank files.
If \code{NULL}, the current working directory (\code{getwd()}) is used. If no GenBank files are found
in the current working directory, the function searches for a subdirectory named "genbank".}

\item{save_output}{Logical, whether to save outputs for each processed file. Default is \code{FALSE}.}
}
\value{
The function assigns the following objects to the global environment:
\itemize{
\item \code{PanDB}: A list of processed GenBank data frames, indexed by filename (excluding extensions).
\item \code{PanDB_df}: A flattened data frame containing all processed data, with an additional column for filenames.
}
}
\description{
This function generates a global list (\code{PanDB}) containing processed data frames for each GenBank file
in a specified directory. Additionally, it creates a flattened data frame (\code{PanDB_df}) for further analysis.
}
\details{
The function processes GenBank files in the specified directory by:
\itemize{
\item Identifying files with extensions \code{.gbk}, \code{.gb}, or \code{.gbff}.
\item Processing each file using the \code{Genbank_organizer} function.
\item Storing the processed data frames in a global list (\code{PanDB}) with filenames as keys.
\item Flattening the list into a single data frame (\code{PanDB_df}) for convenience.
}

The function checks if the specified directory exists and contains valid GenBank files. If not, it stops with an error.
Errors during file processing are logged, and the function continues processing the remaining files.
}
\examples{
# Example usage:
PanDB_generator(directory = "path/to/genbank/files")

# Automatically use the current working directory:
PanDB_generator()

}
\seealso{
\code{\link{Genbank_organizer}}
}
