% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_BPGA_table.R
\name{process_BPGA_table}
\alias{process_BPGA_table}
\title{Process Table}
\usage{
process_BPGA_table(u_cluster = NULL, PanDB_df = NULL, Genome_summary_df = NULL)
}
\arguments{
\item{u_cluster}{A data frame containing cluster information. If \code{NULL}, the function will try
to retrieve \code{u_cluster} from the global environment.}

\item{PanDB_df}{A data frame containing processed PanDB information. If \code{NULL}, the function will
attempt to retrieve \code{PanDB_df} from the global environment.}

\item{Genome_summary_df}{A data frame containing genome summary information. If \code{NULL}, the function
will attempt to retrieve \code{Genome_summary_df} from the global environment.}
}
\value{
The function does not return a value directly but assigns the processed \code{PanDB_df}
to the global environment.
}
\description{
This function processes and filters a final data table by merging input data frames,
performing filtering and validation steps, and saving the result in the global environment as \code{PanDB_df}.
}
\details{
The function:
\itemize{
\item Merges \code{u_cluster} with \code{PanDB_df}.
\item Appends genome summary data from \code{Genome_summary_df}.
\item Filters and removes duplicated or unmatched sequences based on cluster and genome data.
\item Performs validation to ensure all processed files are included.
\item Saves the final processed table as \code{PanDB_df} in the global environment.
}
}
\examples{
# Example usage
process_final_table(u_cluster = my_cluster, PanDB_df = my_PanDB_df, Genome_summary_df = my_genome_summary_df)

# Automatically fetch objects from the global environment
process_final_table()

}
\seealso{
\code{\link{Genbank_organizer_batch}}, \code{\link{combine_gb_info}}
}
